refactor: reduce redundancy by move auxiliary functions to different file
