+++ packages/consumption/src/modules/requests/outgoing/OutgoingRequestsController.ts
@@ -1,19 +1,5 @@
 import { EventBus } from "@js-soft/ts-utils";
-import {
-    CreateAttributeRequestItem,
-    DeleteAttributeRequestItem,
-    ProposeAttributeRequestItem,
-    ReadAttributeRequestItem,
-    RelationshipAttribute,
-    RelationshipAttributeQuery,
-    RelationshipTemplateContent,
-    Request,
-    RequestItem,
-    RequestItemGroup,
-    Response,
-    ResponseItem,
-    ResponseItemGroup
-} from "@nmshd/content";
+import { DeleteAttributeRequestItem, RelationshipTemplateContent, Request, RequestItem, RequestItemGroup, Response, ResponseItem, ResponseItemGroup } from "@nmshd/content";
 import { CoreAddress, CoreDate, CoreId, ICoreId } from "@nmshd/core-types";
 import { Message, PeerDeletionStatus, Relationship, RelationshipStatus, RelationshipTemplate, SynchronizedCollection, TransportCoreErrors } from "@nmshd/transport";
 import { ConsumptionBaseController } from "../../../consumption/ConsumptionBaseController";
@@ -29,6 +15,7 @@ import { RequestItemProcessorRegistry } from "../itemProcessors/RequestItemProce
 import { LocalRequest, LocalRequestSource } from "../local/LocalRequest";
 import { LocalRequestStatus } from "../local/LocalRequestStatus";
 import { LocalResponse } from "../local/LocalResponse";
+import { validateKeyUniquenessOfRelationshipAttributesWithinOutgoingRequest } from "../utility/validateRelationshipAttributesWithinRequest";
 import { CompleteOutgoingRequestParameters, ICompleteOutgoingRequestParameters } from "./completeOutgoingRequest/CompleteOutgoingRequestParameters";
 import {
     CreateAndCompleteOutgoingRequestFromRelationshipTemplateResponseParameters,
@@ -37,12 +24,6 @@ import {
 import { CanCreateOutgoingRequestParameters, ICanCreateOutgoingRequestParameters } from "./createOutgoingRequest/CanCreateOutgoingRequestParameters";
 import { CreateOutgoingRequestParameters, ICreateOutgoingRequestParameters } from "./createOutgoingRequest/CreateOutgoingRequestParameters";
 import { ISentOutgoingRequestParameters, SentOutgoingRequestParameters } from "./sentOutgoingRequest/SentOutgoingRequestParameters";
-
-interface RelationshipAttributeFragment {
-    owner: string;
-    key: string;
-    value: { "@type": string };
-}
 
 export class OutgoingRequestsController extends ConsumptionBaseController {
     public constructor(
@@ -99,7 +80,7 @@ export class OutgoingRequestsController extends ConsumptionBaseController {
         const innerResults = await this.canCreateItems(parsedParams.content, parsedParams.peer);
         const result = ValidationResult.fromItems(innerResults);
 
-        const keyUniquenessValidationResult = OutgoingRequestsController.validateKeyUniquenessOfRelationshipAttributesWithinRequest(parsedParams.content.items, parsedParams.peer);
+        const keyUniquenessValidationResult = validateKeyUniquenessOfRelationshipAttributesWithinOutgoingRequest(parsedParams.content.items, parsedParams.peer);
         if (keyUniquenessValidationResult.isError() && result.isSuccess()) {
             return keyUniquenessValidationResult;
         }
@@ -139,94 +120,6 @@ export class OutgoingRequestsController extends ConsumptionBaseController {
         const result = ValidationResult.fromItems(innerResults);
 
         return result;
-    }
-
-    private static validateKeyUniquenessOfRelationshipAttributesWithinRequest(items: (RequestItem | RequestItemGroup)[], recipient?: CoreAddress) {
-        const fragmentsOfMustBeAcceptedItemsOfRequest: RelationshipAttributeFragment[] = [];
-        for (const item of items) {
-            if (item instanceof RequestItemGroup) {
-                const fragmentsOfMustBeAcceptedItemsOfGroup = OutgoingRequestsController.extractRelationshipAttributeFragmentsFromMustBeAcceptedItemsOfGroup(item, recipient);
-                if (fragmentsOfMustBeAcceptedItemsOfGroup) fragmentsOfMustBeAcceptedItemsOfRequest.push(...fragmentsOfMustBeAcceptedItemsOfGroup);
-            } else {
-                const fragmentOfMustBeAcceptedRequestItem = OutgoingRequestsController.extractRelationshipAttributeFragmentFromMustBeAcceptedRequestItem(item, recipient);
-                if (fragmentOfMustBeAcceptedRequestItem) fragmentsOfMustBeAcceptedItemsOfRequest.push(fragmentOfMustBeAcceptedRequestItem);
-            }
-        }
-
-        if (OutgoingRequestsController.containsDuplicateRelationshipAttributeFragments(fragmentsOfMustBeAcceptedItemsOfRequest)) {
-            return ValidationResult.error(
-                ConsumptionCoreErrors.requests.violatedKeyUniquenessOfRelationshipAttributes(
-                    "The Request cannot be created because its acceptance would lead to the creation of more than one RelationshipAttribute in the context of this Relationship with the same key, owner and value type."
-                )
-            );
-        }
-
-        return ValidationResult.success();
-    }
-
-    private static extractRelationshipAttributeFragmentsFromMustBeAcceptedItemsOfGroup(requestItemGroup: RequestItemGroup, recipient?: CoreAddress) {
-        const fragmentsOfMustBeAcceptedItemsOfGroup: RelationshipAttributeFragment[] = [];
-
-        for (const item of requestItemGroup.items) {
-            if (item instanceof RequestItemGroup) {
-                const fragments = OutgoingRequestsController.extractRelationshipAttributeFragmentsFromMustBeAcceptedItemsOfGroup(item, recipient);
-                if (fragments) fragmentsOfMustBeAcceptedItemsOfGroup.push(...fragments);
-            } else {
-                const fragment = OutgoingRequestsController.extractRelationshipAttributeFragmentFromMustBeAcceptedRequestItem(item, recipient);
-                if (fragment) fragmentsOfMustBeAcceptedItemsOfGroup.push(fragment);
-            }
-        }
-
-        if (fragmentsOfMustBeAcceptedItemsOfGroup.length !== 0) return fragmentsOfMustBeAcceptedItemsOfGroup;
-
-        return;
-    }
-
-    private static extractRelationshipAttributeFragmentFromMustBeAcceptedRequestItem(requestItem: RequestItem, recipient?: CoreAddress) {
-        if (requestItem.mustBeAccepted) {
-            return OutgoingRequestsController.extractRelationshipAttributeFragmentFromRequestItem(requestItem, recipient);
-        }
-
-        return;
-    }
-
-    private static extractRelationshipAttributeFragmentFromRequestItem(requestItem: RequestItem, recipient?: CoreAddress) {
-        if (requestItem instanceof CreateAttributeRequestItem && requestItem.attribute instanceof RelationshipAttribute) {
-            const ownerIsEmptyString = requestItem.attribute.owner.toString() === "";
-            return {
-                owner: ownerIsEmptyString && recipient ? recipient.toString() : requestItem.attribute.owner.toString(),
-                key: requestItem.attribute.key,
-                value: { "@type": requestItem.attribute.value.toJSON()["@type"] }
-            };
-        } else if (
-            (requestItem instanceof ReadAttributeRequestItem || requestItem instanceof ProposeAttributeRequestItem) &&
-            requestItem.query instanceof RelationshipAttributeQuery
-        ) {
-            const ownerIsEmptyString = requestItem.query.owner.toString() === "";
-            return {
-                owner: ownerIsEmptyString && recipient ? recipient.toString() : requestItem.query.owner.toString(),
-                key: requestItem.query.key,
-                value: { "@type": requestItem.query.attributeCreationHints.valueType }
-            };
-        }
-
-        return;
-    }
-
-    private static containsDuplicateRelationshipAttributeFragments(fragments: RelationshipAttributeFragment[]) {
-        const seenIdentifier = new Set<string>();
-
-        for (const fragment of fragments) {
-            const identifierOfFragment = JSON.stringify(fragment);
-
-            if (seenIdentifier.has(identifierOfFragment)) {
-                return true;
-            }
-
-            seenIdentifier.add(identifierOfFragment);
-        }
-
-        return false;
     }
 
     public async create(params: ICreateOutgoingRequestParameters): Promise<LocalRequest> {
